+++ Communication/Packets/Incoming/Catalog/ShopTargetedOfferViewedEvent.cs
@@ -1,8 +1,17 @@
+using Plus.HabboHotel.Catalog;
 using Plus.HabboHotel.GameClients;
 
 namespace Plus.Communication.Packets.Incoming.Catalog;
 
 internal class ShopTargetedOfferViewedEvent : IPacketEvent
 {
-    public Task Parse(GameClient session, IIncomingPacket packet) => throw new NotImplementedException();
+    private readonly ITargetedOfferService _targetedOfferService;
+
+    public ShopTargetedOfferViewedEvent(ITargetedOfferService targetedOfferService) => _targetedOfferService = targetedOfferService;
+
+    public Task Parse(GameClient session, IIncomingPacket packet)
+    {
+        int? offerId = packet.HasDataRemaining() ? packet.ReadInt() : null;
+        return _targetedOfferService.MarkViewed(session, offerId);
+    }
 }